requirements.txt
# Requirements
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
# mysqlclient==2.2.5
pillow==11.0.0
PyJWT==2.9.0
sendgrid==6.11.0
python_dotenv==1.0.1